+++ Sources/BFKit/Apple/UIKit/UIImageExtension.swift
@@ -616,6 +616,23 @@ public extension UIImage {
     }
     
     #if os(iOS)
+        /// Creates an image from an UIView.
+        ///
+        /// - Parameter view: UIView.
+        public convenience init?(view: UIView) {
+            UIGraphicsBeginImageContext(view.frame.size)
+            guard let context = UIGraphicsGetCurrentContext() else {
+                return nil
+            }
+            view.layer.render(in: context)
+            let image = UIGraphicsGetImageFromCurrentImageContext()
+            UIGraphicsEndImageContext()
+            guard let cgImage = image?.cgImage else {
+                return nil
+            }
+            self.init(cgImage: cgImage)
+        }
+    
         /// Rotate the image to the given radians.
         ///
         /// - Parameter radians: Radians to rotate to